Moved InlineDialect/Parser to plugins
