+++ frictionless/dialects.py
@@ -92,82 +92,6 @@ class Dialect(Metadata):
     }
 
 
-class InlineDialect(Dialect):
-    """Inline dialect representation
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless import dialects`
-
-    Parameters:
-        descriptor? (str|dict): descriptor
-        keys? (str[]): a list of strings to use as data keys
-        keyed? (bool): whether data rows are keyed
-
-    Raises:
-        FrictionlessException: raise any error that occurs during the process
-
-    """
-
-    def __init__(
-        self,
-        descriptor=None,
-        *,
-        keys=None,
-        keyed=None,
-        header=None,
-        header_rows=None,
-        header_join=None,
-        header_case=None,
-    ):
-        self.setinitial("keys", keys)
-        self.setinitial("keyed", keyed)
-        super().__init__(
-            descriptor=descriptor,
-            header=header,
-            header_rows=header_rows,
-            header_join=header_join,
-            header_case=header_case,
-        )
-
-    @Metadata.property
-    def keys(self):
-        """
-        Returns:
-            str[]?: keys
-        """
-        return self.get("keys")
-
-    @Metadata.property
-    def keyed(self):
-        """
-        Returns:
-            bool: keyed
-        """
-        return self.get("keyed", False)
-
-    # Expand
-
-    def expand(self):
-        """Expand metadata"""
-        self.setdefault("keyed", self.keyed)
-
-    # Metadata
-
-    metadata_profile = {  # type: ignore
-        "type": "object",
-        "additionalProperties": False,
-        "properties": {
-            "keys": {"type": "array"},
-            "keyed": {"type": "boolean"},
-            "header": {"type": "boolean"},
-            "headerRows": {"type": "array", "items": {"type": "number"}},
-            "headerJoin": {"type": "string"},
-            "headerCase": {"type": "boolean"},
-        },
-    }
-
-
 class JsonDialect(Dialect):
     """Json dialect representation
