+++ frictionless/parsers/__init__.py
@@ -1,2 +1,1 @@
-from .inline import InlineParser
 from .json import JsonParser, JsonlParser
